+++ TaskbarNotifier.py
@@ -355,6 +355,11 @@ class MainWindow(QWidget):
         show_action.triggered.connect(self.__on_show)
         tray_menu.addAction(show_action)
 
+        self.tray_enable_disable_action = QAction("Disable", self)
+        self.tray_enable_disable_action.setIcon(self.style().standardIcon(QStyle.SP_MediaPause))
+        self.tray_enable_disable_action.triggered.connect(self.__on_enable_disable)
+        tray_menu.addAction(self.tray_enable_disable_action)
+
         about_action = QAction("About", self)
         about_action.setIcon(self.style().standardIcon(QStyle.SP_FileDialogInfoView))
         about_action.triggered.connect(self.__on_about)
@@ -498,6 +503,23 @@ class MainWindow(QWidget):
         self.timer_polling.stop()
         self.show()
         self.activateWindow()
+
+    def __on_enable_disable(self) -> None:
+        """
+        Event handler for the tray enable/disable action.
+        """
+
+        if self.timer_polling.isActive():
+            self.applications_on_toast = []
+            self.tray_icon.setIcon(QIcon(":/Disabled.png"))
+            self.tray_enable_disable_action.setText("Enable")
+            self.tray_enable_disable_action.setIcon(self.style().standardIcon(QStyle.SP_MediaPlay))
+            self.timer_polling.stop()
+        else:
+            self.tray_icon.setIcon(QIcon(":/Grey.png"))
+            self.tray_enable_disable_action.setText("Disable")
+            self.tray_enable_disable_action.setIcon(self.style().standardIcon(QStyle.SP_MediaPause))
+            self.timer_polling.start(self.TIMER_INVTERVAL_POLLING_MS)
 
     @staticmethod
     def __on_about() -> None: