+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file. The format 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add enable/disable control to tray icon context menu.
+
 ### Changed
 - Allow Taskbar Notifier to run on systems without an installed Git client